:skip_optional_callbacks option to `defmock` skips optional callbacks
This is useful if you have code that checks if a given function is
implemented and only calls it in that case (or has completely separate
behavior in the not-implemented case). By default all optional callbacks
are implemented as functions.

Note: in the tests I didn't name all the defined mocks `MyMock` because
I received warnings like:

    warning: redefining module MyMock (current version defined in memory)
      lib/mox.ex:316