Highlight the utility of stub_with/2
I think it's most useful for when we can't pass the mock to the
system under test directly/we have to configure it in the test
environment. Hence I added it right underneath that section.

Not sure if that's the best choice or if a separate section
would be better. Happy for feedback!

Follow up to #76